refactor: extract TranslateTagsUseCase into its own module
Each use case should have its own file. Moved TranslateTagsUseCase from translate_article.py to translate_tags.py.
