+++ src/modules/intelligence/application/use_cases/__init__.py
@@ -1,6 +1,7 @@
 from .analyze_article import AnalyzeArticleUseCase
 from .analysis_result import AnalysisResult
-from .translate_article import TranslateArticleUseCase, TranslateTagsUseCase
+from .translate_article import TranslateArticleUseCase
+from .translate_tags import TranslateTagsUseCase
 
 __all__ = [
     'AnalyzeArticleUseCase',
